[PORT FROM MAIN][Audio VOIP] ignore NORMAL mode
BZ: 63505

Root Cause of issue is linked to timing problem to stop
the threads. Because the issue is not reproducible on
reference phone, one of possible fixes is to find different
between our code sources and reference phones code sources.
According comments in code source, there is issue on
Samsung tuna and herring boards (i.e. on reference phones).
They are defined a workaround and this patch purpose extended
the Samsung workaround to Intel products..

Orig-Change-Id: I73b6da487138a636d664be004307e819b2bbe768

Change-Id: Iee394ba3b672289d8a7a74fea94d96bf0e94fee2
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/123592
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>